Add implementation for GetMe
