+++ api/auth.go
@@ -3,7 +3,9 @@ package api
 import (
 	"connectrpc.com/connect"
 	"context"
+	"errors"
 	"fmt"
+	"github.com/jackc/pgx/v5"
 	"github.com/strideynet/bsky-furry-feed/bluesky"
 	v1 "github.com/strideynet/bsky-furry-feed/proto/bff/v1"
 	"go.uber.org/zap"
@@ -30,6 +32,8 @@ func BSkyTokenValidator(pdsHost string) func(ctx context.Context, token string) 
 // authenticatedUserPermissions are granted to any user who is authenticated.
 var authenticatedUserPermissions = []string{
 	"/bff.v1.ModerationService/Ping",
+	"/bff.v1.UserService/GetMe",
+	"/bff.v1.UserService/JoinApprovalQueue",
 }
 
 var moderatorPermissions = []string{
@@ -65,7 +69,14 @@ type AuthEngine struct {
 }
 
 type authContext struct {
+	// DID is the did extracted from the token supplied by the user.
 	DID string
+	// Actor is the actor fetched from the database during authz/authn. This
+	// should be used carefully, and if necessary the actor should be fetched
+	// again within a transaction if mutation is occurring.
+	//
+	// This will be nil if the actor does not exist.
+	Actor *v1.Actor
 }
 
 // TODO: Allow a authOpts to be passed in with a description of attempted
@@ -93,8 +104,13 @@ func (a *AuthEngine) auth(ctx context.Context, req connect.AnyRequest) (*authCon
 	// Find the actor in the database so we know their roles and status to
 	// be able to evaluate authz.
 	actor, err := a.ActorGetter.GetActorByDID(ctx, did)
-	if err != nil {
+	if err != nil && !errors.Is(err, pgx.ErrNoRows) {
 		return nil, fmt.Errorf("fetching actor for token: %w", err)
+	}
+
+	actorRoles := []string{}
+	if actor != nil {
+		actorRoles = actor.Roles
 	}
 
 	// Use a map of string to bool as a quasi set.
@@ -105,7 +121,7 @@ func (a *AuthEngine) auth(ctx context.Context, req connect.AnyRequest) (*authCon
 		permissions[permission] = true
 	}
 	// Now we grant them all the permissions from their roles
-	for _, role := range actor.Roles {
+	for _, role := range actorRoles {
 		rolePerms, ok := roleToPermissions[role]
 		if !ok {
 			// Gracefully handle an unrecognized role
@@ -131,6 +147,7 @@ func (a *AuthEngine) auth(ctx context.Context, req connect.AnyRequest) (*authCon
 	}
 
 	return &authContext{
-		DID: actor.Did,
+		DID:   actor.Did,
+		Actor: actor,
 	}, nil
 }
