--- buf.work.yaml
@@ -1,3 +1,0 @@
-version: v1
-directories:
-  - proto
