+++ proto/bff/v1/user_service.proto
@@ -2,6 +2,7 @@ syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
 import "google/protobuf/any.proto";
+import "bff/v1/types.proto";
 
 package bff.v1;
 
@@ -13,7 +14,9 @@ service UserService {
 }
 
 message GetMeRequest {}
-message GetMeResponse {}
+message GetMeResponse {
+  bff.v1.Actor Actor = 1;
+}
 
 message JoinApprovalQueueRequest {}
 message JoinApprovalQueueResponse {}
