+++ web/proto/bff/v1/types_pb.js
@@ -1,0 +1,36 @@
+// @generated by protoc-gen-es v1.2.0 with parameter "target=js+dts"
+// @generated from file bff/v1/types.proto (package bff.v1, syntax proto3)
+/* eslint-disable */
+// @ts-nocheck
+
+import { proto3, Timestamp } from "@bufbuild/protobuf";
+
+/**
+ * @generated from enum bff.v1.ActorStatus
+ */
+export const ActorStatus = proto3.makeEnum(
+  "bff.v1.ActorStatus",
+  [
+    {no: 0, name: "ACTOR_STATUS_UNSPECIFIED", localName: "UNSPECIFIED"},
+    {no: 1, name: "ACTOR_STATUS_PENDING", localName: "PENDING"},
+    {no: 2, name: "ACTOR_STATUS_APPROVED", localName: "APPROVED"},
+    {no: 3, name: "ACTOR_STATUS_BANNED", localName: "BANNED"},
+    {no: 4, name: "ACTOR_STATUS_NONE", localName: "NONE"},
+  ],
+);
+
+/**
+ * @generated from message bff.v1.Actor
+ */
+export const Actor = proto3.makeMessageType(
+  "bff.v1.Actor",
+  () => [
+    { no: 1, name: "did", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 3, name: "is_artist", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
+    { no: 4, name: "comment", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 5, name: "status", kind: "enum", T: proto3.getEnumType(ActorStatus) },
+    { no: 6, name: "created_at", kind: "message", T: Timestamp },
+    { no: 7, name: "roles", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
+  ],
+);
+
